split xwayland-view-base out of xwayland.cpp
