+++ src/view/xwayland/xwayland-helpers.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "config.h"
 #include <optional>
 #include <string>
